// resources/admin_ui/src/app/interfaces/laravel/paginated.ts
import { Link } from './link';

export interface Paginated<T> {
    current_page: number;
    data: Array<T>;
    first_page_url: string;
    from: number | null;
    last_page: number;
    last_page_url: string;
    links: Array<Link>,
    next_page_url: string | null;
    path: string;
    per_page: number;
    prev_page_url: string | null;
    to: number | null;
    total: number;
}